requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.2
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
setuptools==70.2.0
sqlparse==0.5.0
tzdata==2024.1
wheel==0.43.0
